+++ lib/biotcm/databases.rb
@@ -3,4 +3,5 @@
 # Namespace of all classes who interactive with databases
 # directly or indirectly
 module BioTCM::Databases
+  autoload(:HGNC, "biotcm/databases/hgnc")
 end