+++ test/test-helper.rb
@@ -6,3 +6,6 @@ unless $:.include?(File.expand_path('../../lib', __FILE__))
   $:.unshift(File.expand_path('../../lib', __FILE__))
 end
 require 'biotcm'
+
+# Suppress output from screen_logger
+BioTCM.log.screen_logger.level = Logger::UNKNOWN